refactor: extract agent event contract to core, break compiler->agent dependency
src/compiler/mcp-replay.ts imported normalizeAgentEvent from agent/host.ts -
a cross-layer reach (io -> agent) flagged by architecture.yml.

Move the normalized agent event contract (event/usage types + normalization
pipeline + error-message helpers) into src/core/agent-events.ts. host.ts
imports the types it still uses and re-exports the contract for agent-layer
consumers (runner, progress, execution-receipts, tests).

Verified: tsc --noEmit clean, 486 vitest tests pass, tsc build clean,
architecture contract checker reports 0 violations.
